M  src/meta/embedded/src/lib.rs
