+++ src/meta/raft-store/src/mem_meta.rs
@@ -19,7 +19,6 @@ use std::sync::Arc;
 use databend_common_meta_kvapi::kvapi::KVApi;
 use databend_common_meta_kvapi::kvapi::KVStream;
 use databend_common_meta_types::protobuf::StreamItem;
-use databend_common_meta_types::sys_data::SysData;
 use databend_common_meta_types::AppliedState;
 use databend_common_meta_types::Change;
 use databend_common_meta_types::CmdContext;
@@ -35,46 +34,8 @@ use log::debug;
 use tokio::sync::Mutex;
 
 use crate::applier::Applier;
-use crate::leveled_store::level::Level;
-use crate::state_machine::ExpireKey;
-use crate::state_machine::StateMachineSubscriber;
-use crate::state_machine_api::StateMachineApi;
+use crate::mem_state_machine::MemStateMachine;
 use crate::state_machine_api_ext::StateMachineApiExt;
-
-/// A pure in-memory state machine as mock for testing.
-#[derive(Debug, Default)]
-pub struct MemStateMachine {
-    level: Level,
-    expire_cursor: ExpireKey,
-}
-
-impl StateMachineApi for MemStateMachine {
-    type Map = Level;
-
-    fn get_expire_cursor(&self) -> ExpireKey {
-        self.expire_cursor
-    }
-
-    fn set_expire_cursor(&mut self, cursor: ExpireKey) {
-        self.expire_cursor = cursor;
-    }
-
-    fn map_ref(&self) -> &Self::Map {
-        &self.level
-    }
-
-    fn map_mut(&mut self) -> &mut Self::Map {
-        &mut self.level
-    }
-
-    fn sys_data_mut(&mut self) -> &mut SysData {
-        &mut self.level.sys_data
-    }
-
-    fn get_subscriber(&self) -> Option<&dyn StateMachineSubscriber> {
-        None
-    }
-}
 
 #[derive(Clone, Default)]
 pub struct MemMeta {
