+++ src/query/management/tests/it/cluster.rs
@@ -18,8 +18,8 @@ use std::time::Duration;
 use databend_common_base::base::tokio;
 use databend_common_exception::Result;
 use databend_common_management::*;
-use databend_common_meta_embedded::MemMeta;
 use databend_common_meta_kvapi::kvapi::KVApi;
+use databend_common_meta_raft_store::mem_meta::MemMeta;
 use databend_common_meta_store::MetaStore;
 use databend_common_meta_types::seq_value::SeqV;
 use databend_common_meta_types::MatchSeq;
